Implemented Topological with generic algorithm schema
