+++ sampo/scheduler/generic.py
@@ -1,8 +1,10 @@
 from typing import Type, List, Callable, Iterable
 
 from sampo.scheduler.base import Scheduler, SchedulerType
+from sampo.scheduler.heft.time_computaion import calculate_working_time_cascade
 from sampo.scheduler.resource.base import ResourceOptimizer
 from sampo.scheduler.timeline.base import Timeline
+from sampo.scheduler.utils.multi_contractor import run_contractor_search, get_worker_borders
 from sampo.schemas.contractor import Contractor, get_worker_contractor_pool, WorkerContractorPool
 from sampo.schemas.graph import WorkGraph, GraphNode
 from sampo.schemas.resources import Worker
@@ -12,6 +14,12 @@ from sampo.schemas.scheduled_work import ScheduledWork
 from sampo.schemas.time import Time
 from sampo.schemas.time_estimator import WorkTimeEstimator
 from sampo.utilities.validation import validate_schedule
+
+
+def get_finish_time_default(node, worker_team, node2swork, assigned_parent_time, timeline, work_estimator):
+    return timeline.find_min_start_time(node, worker_team, node2swork,
+                                        assigned_parent_time, work_estimator) \
+        + calculate_working_time_cascade(node, worker_team, work_estimator)
 
 
 class GenericScheduler(Scheduler):
@@ -29,6 +37,43 @@ class GenericScheduler(Scheduler):
         self._timeline_type = timeline_type
         self.prioritization = prioritization_f
         self.optimize_resources = optimize_resources_f
+
+    def get_default_res_opt_function(self, get_finish_time = get_finish_time_default) \
+            -> Callable[[GraphNode, list[Contractor], WorkSpec, WorkerContractorPool,
+                         dict[GraphNode, ScheduledWork], Time, Timeline, WorkTimeEstimator],
+                         tuple[Time, Time, Contractor, list[Worker]]]:
+
+        def optimize_resources_def(node: GraphNode, contractors: list[Contractor], work_spec: WorkSpec,
+                                   worker_pool: WorkerContractorPool, node2swork: dict[GraphNode, ScheduledWork],
+                                   assigned_parent_time: Time, timeline: Timeline, work_estimator: WorkTimeEstimator):
+            def run_with_contractor(contractor: Contractor) -> tuple[Time, Time, List[Worker]]:
+                min_count_worker_team, max_count_worker_team, workers \
+                    = get_worker_borders(worker_pool, contractor, node.work_unit.worker_reqs)
+
+                if len(workers) != len(node.work_unit.worker_reqs):
+                    return Time(0), Time.inf(), []
+
+                workers = [worker.copy() for worker in workers]
+
+                def ft_getter(worker_team):
+                    return get_finish_time(node, worker_team, node2swork, assigned_parent_time, timeline, work_estimator)
+
+
+                # apply worker team spec
+                self.optimize_resources_using_spec(node.work_unit, workers, work_spec,
+                                                   lambda optimize_array: self.resource_optimizer.optimize_resources(
+                                                       worker_pool, workers,
+                                                       optimize_array,
+                                                       min_count_worker_team, max_count_worker_team,
+                                                       ft_getter))
+
+                c_st, c_ft, _ = timeline.find_min_start_time_with_additional(node, workers, node2swork, None,
+                                                                             assigned_parent_time, work_estimator)
+                return c_st, c_ft, workers
+
+            return run_contractor_search(contractors, run_with_contractor)
+
+        return optimize_resources_def
 
     def schedule_with_cache(self, wg: WorkGraph,
                             contractors: List[Contractor],
